+++ lib/genex/chromosome.ex
@@ -18,7 +18,7 @@ defmodule Genex.Chromosome do
   defstruct [:genes, fitness: 0, size: 0, age: 0]
 
   @doc """
-  Creates a binary geneset.
+  Creates a binary encoded geneset.
 
   Returns `Enum.t()`.
 
@@ -27,11 +27,79 @@ defmodule Genex.Chromosome do
     - `opts`: Configuration options.
 
   # Options
+
     - `size`: Size of the chromosomes. Defaults to 32.
   """
   def binary(opts \\ []) do
     size = Keyword.get(opts, :size, 32)
     for _ <- 0..(size - 1), do: Enum.random(0..1)
+  end
+
+  @doc """
+  Creates an integer value encoded geneset.
+
+  Returns `Enum.t()`.
+
+  # Parameters
+
+    - `opts`: Configuration options.
+
+  # Options
+
+    - `size`: Size of the chromosome. Defaults to 32.
+    - `min`: Minimum number of distribution. Defaults to 0.
+    - `max`: Maximum number of distribution. Defaults to 10.
+  """
+  def integer_value(opts \\ []) do
+    size = Keyword.get(opts, :size, 32)
+    low = Keyword.get(opts, :min, 0)
+    high = Keyword.get(opts, :max, 0)
+    for _ <- 0..(size - 1), do: Enum.random(low..high)
+  end
+
+  @doc """
+  Creates a permutation encoded geneset.
+
+  Returns `Enum.t()`.
+
+  # Parameters
+
+    - `opts`: Configuration options.
+
+  # Options
+
+    - `min`: Minimum number of distribution. Defaults to 1.
+    - `max`: Maximum number of distribution. Defaults to 10.
+  """
+  def permutation(opts \\ []) do
+    low = Keyword.get(opts, :min, 1)
+    high = Keyword.get(opts, :max, 10)
+
+    low..high
+    |> Enum.shuffle()
+  end
+
+  @doc """
+  Creates an alphabetically encoded geneset.
+
+  Returns `Enum.t()`.
+
+  # Parameters
+
+    - `opts`: Configuration options.
+
+  # Options
+
+    - `size`: Size of the chromosome. Defaults to 10.
+  """
+  def alphabetic(opts \\ []) do
+    size = Keyword.get(opts, :size, 10)
+
+    ?a..?z
+    |> Enum.to_list()
+    |> Enum.shuffle()
+    |> Enum.split(size)
+    |> elem(0)
   end
 
   defimpl String.Chars, for: __MODULE__ do